we should really make transferFactor a double.

+++ src/main/java/appeng/tile/misc/CondenserVoidInventory.java
@@ -51,7 +51,7 @@ class CondenserVoidInventory<T extends IAEStack<T>> implements IMEMonitor<T>
 
 		if( input != null )
 		{
-			this.target.addPower( input.getStackSize() / this.channel.transferFactor() );
+			this.target.addPower( input.getStackSize() / (double) this.channel.transferFactor() );
 		}
 		return null;
 	}
